requirements.txt
Flask
Flask-SocketIO
eventlet
opencv-python
bleak
PyYAML
pyserial
numpy
